chore: add meal-plan schema and helpers
