+++ Nutrishop/nutrishop-v2/next.config.mjs
@@ -1,8 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-  experimental: {
-    appDir: true
-  }
-}
+const nextConfig = {}
 
 export default nextConfig
